Allow resizing of Drawer() internal pixmaps
drawer.width is set in Popup when the Popup's width changes.
Unfortunately, the underlying drawing surface on X(pixmap) is not
resized. It doesn't seem like they can be resized after reading the x
proto docs. Cairo to the rescue! Cairo implements a RecordingSurface
that we can use locally while the underlying Drawable may not be
accessible. These changes make the Drawer write to a RecordingSurface
and only copy to the pixmap on Drawer.draw.

This relies on functionality implemented in
https://github.com/tych0/xcffib/pull/114